Added dependency injection, database service, signin/up API and more.
Using Typedi to inject a database service to abstract away database /
sqlite from Graphql resolvers.

Created a user Typedi service that extends the database service (so has
access to the database), where auth related API calls are placed (e.g
getUserByUsername, createUser). Should perhaps be renamed to
authentication service.

Using Apollo CLI Typescript types are generated from the TypeGraphQL
schema (backend is run with --genSchema before frontend build).

Added comma dangle syntax rule to make multi-line lists look cleaner.

Functionality:

Signing in and up is working (though without validation on the
backend).

Authentication is done with the httpOnly cookie authToken, set to a
long random string of characters (Base64), and stored in the database.
The table is then checked to see if the cookie is valid, when a client
queries etc.

Password hashes & salts are stored in a seperate table to not pollute
the User type which could be dangerous. Hashing is done with Scrypt.