Initial Access Control commit, implementing ACL authorization.
An ACL architecture is used for authorization and managing what
permissions users have when accessing certain data (resources, other
users details etc).

The custom AccessControl decorators first argument is the level of
access the method should require from the user (READ, WRITE, or DELETE),
for instance deleteUser() requires DELETE access levels on 'user',
whereas getSongName requires READ access levels on 'resource'.

The decorator takes the supplied target resource (such as another
user) from the called resolver, and then checks whether the user which
called the resolver is permitted to access the resource at the level
which the resolver calls for.

If the decorator is empty, it is simply checked whether the user is
logged in.

Additional changes:

DatabaseService was refactored to extend the better-sqlite3 class in
order to prevent duplication of the service (it was being instantiated
multiple times), a better workaround should possibly be found.

LoggerService renamed to LoggingService, output prettified with colours
on the terminal.